@@ -11,17 +11,17 @@
         "psr-4": { "Demontpx\\ParsedownBundle\\": "tests/" }
     },
     "require": {
-        "php": "^7.1",
-        "symfony/config": "^3.4|^4.0|^5.0",
-        "symfony/http-kernel": "^3.4|^4.0|^5.0",
-        "symfony/dependency-injection": "^3.4|^4.0|^5.0",
+        "php": "^7.4",
+        "symfony/config": "^4.4|^5.0",
+        "symfony/http-kernel": "4.4|^5.0",
+        "symfony/dependency-injection": "^4.4|^5.0",
         "twig/twig": "^2.4|^3.0",
         "erusev/parsedown": "^1.6"
     },
     "require-dev": {
         "phpunit/phpunit": "^8.4",
-        "mockery/mockery": "^1.2",
-        "symfony/yaml": "^3.4|^4.0|^5.0"
+        "mockery/mockery": "^1.3",
+        "symfony/yaml": "^4.4|^5.0"
     },
     "config": {
         "bin-dir": "bin"
